move CdsConfig to CdsLoadBalancerProvider
