+++ xds/src/main/java/io/grpc/xds/CdsLoadBalancerProvider.java
@@ -16,6 +16,8 @@
 
 package io.grpc.xds;
 
+import static com.google.common.base.Preconditions.checkArgument;
+
 import io.grpc.Internal;
 import io.grpc.LoadBalancer;
 import io.grpc.LoadBalancer.Helper;
@@ -23,8 +25,8 @@ import io.grpc.LoadBalancerProvider;
 import io.grpc.NameResolver.ConfigOrError;
 import io.grpc.Status;
 import io.grpc.internal.JsonUtil;
-import io.grpc.xds.CdsLoadBalancer.CdsConfig;
 import java.util.Map;
+import java.util.Objects;
 
 @Internal
 public class CdsLoadBalancerProvider extends LoadBalancerProvider {
@@ -70,6 +72,39 @@ public class CdsLoadBalancerProvider extends LoadBalancerProvider {
     } catch (RuntimeException e) {
       return ConfigOrError.fromError(
           Status.UNKNOWN.withDescription("Failed to parse config " + e.getMessage()).withCause(e));
+    }
+  }
+
+  /**
+   * Represents a successfully parsed and validated LoadBalancingConfig for CDS.
+   */
+  static final class CdsConfig {
+
+    /**
+     * Name of cluster to query CDS for.
+     */
+    final String name;
+
+    CdsConfig(String name) {
+      checkArgument(name != null && !name.isEmpty(), "name is null or empty");
+      this.name = name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+      if (this == o) {
+        return true;
+      }
+      if (o == null || getClass() != o.getClass()) {
+        return false;
+      }
+      CdsConfig cdsConfig = (CdsConfig) o;
+      return Objects.equals(name, cdsConfig.name);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(name);
     }
   }
 }
